+++ src/iteration_utilities/_iteration_utilities/helpercompat.h
@@ -23,23 +23,6 @@ extern "C" {
   #endif
 #endif
 
-#if PYIU_PYPY
-  // Taken from PyObject_HEAD_INIT implementation (it's a bit hacky...)
-  #define PYIU_CREATE_SINGLETON_INSTANCE(type) { 1, 0, &type }
-#else
-  #if PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION < 12
-    #define PYIU_CREATE_SINGLETON_INSTANCE(type) { _PyObject_EXTRA_INIT 1, &type }
-  #elif PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION == 12
-    #define PYIU_CREATE_SINGLETON_INSTANCE(type) { _PyObject_EXTRA_INIT { _Py_IMMORTAL_REFCNT }, &type }
-  #elif PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION == 13
-    // It seems like they will remove the _PyObject_EXTRA_INIT (not present on master of CPython anymore)
-    // it could still be subject to change but compare to the definition for _Py_NotImplementedStruct or _Py_NoneStruct
-    #define PYIU_CREATE_SINGLETON_INSTANCE(type) { { _Py_IMMORTAL_REFCNT }, &type }
-  #else
-    #define PYIU_CREATE_SINGLETON_INSTANCE(type) { { _Py_IMMORTAL_INITIAL_REFCNT }, &type }
-  #endif
-#endif
-
 #define PyIU_USE_FASTCALL (PYIU_CPYTHON && PY_MAJOR_VERSION == 3 && (PY_MINOR_VERSION == 6 || PY_MINOR_VERSION == 7))
 #define PyIU_USE_UNDERSCORE_VECTORCALL (PYIU_CPYTHON && PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION == 8)
 #define PyIU_USE_VECTORCALL (PYIU_CPYTHON && ((PY_MAJOR_VERSION == 3 && PY_MINOR_VERSION >= 8) || PY_MAJOR_VERSION >= 4))